Add: Heel Heights functionalities and APIs

+++ database/migrations/2024_11_13_030042_create_products_heel_heights_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products_heel_heights', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->integer('heel_heights');
+            $table->decimal('heel_heights', 5, 2);
         });
     }
